+++ indent.js
@@ -2,39 +2,55 @@
 // (using the MINIMAL LISP READER)
 //
 // Fix the indentation of every line using very simple rules.
+//
+// ENFORCE indentation to imply nesting.
+// ENFORCE vertical alignment of siblings.
+// ALLOW 1-space or 2-space or Arg-alignment (chosen by first sibling).
 
-import { UINT_NULL, peek, isOpenParen, readText } from "./read.js";
+import { peek, isOpenParen, readText } from "./read.js";
 
 //------------------------------------------------------------------------------
 // Indentation correction
 //------------------------------------------------------------------------------
 
 function getOpenerIndentSize(state, opener) {
-  if (opener.ch === "[") {
-    return 1;
-  } else if (opener.ch === "{") {
-    return 1;
-  } else if (opener.ch === "(") {
-    const line = state.lines[opener.lineNo];
-    const token = getToken(line, opener.x + 1);
-    if (isSymbol(token.str)) {
-      const argX = seek(line, token.end, ch => ch !== " ");
-      const argExists = argX < line.length && line[argX] !== ";";
-      const isAligned = argX === state.x;
-      if (argExists && isAligned) {
-        const n = argX - opener.x;
-        return n;
-      }
-      return 2;
-    }
-    return 1;
-  }
+  // get all indent sizes
+  // (foo bar baz quxx
+  //  ||            <-- standard 1 and 2 space
+  //      |   |   | <-- arg stops > 2
+  //  )
+  const argStops = opener.children
+    .map(tok => tok.x - opener.x)
+    .filter(size => 2 < size);
+  const sizes = [1, 2, ...argStops];
+
+  // cap indentation size (i.e. parinfer)
+  // ((foo)
+  //  ||
+  //   x      <-- remove two-space to prevent indentation inside (foo)
+  //  )
+  const [a, b] = opener.children;
+  const collFirst = a && isOpenParen(a.ch);
+  const preventCollIndent = !b || b.ch === ";";
+  const maxSize = collFirst && preventCollIndent ? a.x - opener.x : Infinity;
+
+  // return current size if valid
+  const validSizes = sizes.filter(size => size <= maxSize);
+  const currSize = state.x - opener.x;
+  if (validSizes.includes(currSize)) return currSize;
+
+  // revert to default if valid
+  const defaultSize = { "(": 2, "{": 1, "[": 1 }[opener.ch];
+  if (validSizes.includes(defaultSize)) return defaultSize;
+
+  // fallback to 1
+  return 1;
 }
 
 function getCorrectIndentX(state) {
   if (state.parenStack.length === 0) return 0;
   const opener = peek(state.parenStack, 0);
-  if (opener.childIndentX === UINT_NULL) {
+  if (!opener.childIndentX) {
     const indentSize = getOpenerIndentSize(state, opener);
     opener.childIndentX = opener.x + opener.indentDelta + indentSize;
   }
@@ -66,7 +82,6 @@ function onIndent(state) {
 
 function onOpener(state, opener) {
   opener.indentDelta = state.indentDelta;
-  opener.childIndentX = UINT_NULL;
 }
 
 const hooks = {
@@ -75,38 +90,6 @@ const hooks = {
   onIndent,
   onOpener
 };
-
-//------------------------------------------------------------------------------
-// Extra reader code
-//------------------------------------------------------------------------------
-
-function seek(line, startX, pred) {
-  for (let x = startX; x < line.length; x++) {
-    if (pred(line[x])) return x;
-  }
-  return line.length;
-}
-
-function getToken(line, x) {
-  const start = seek(line, x, ch => ch !== " ");
-  const end = seek(line, start + 1, ch => ' \\"({[;'.includes(ch));
-  const str = line.slice(start, end);
-  return { start, end, str };
-}
-
-function isSymbol(str) {
-  if (str === "") return false;
-  if (str[0].match(/\d/)) return false;
-  if (str[0] === ":") return false;
-  if (str[0] === "#") return false;
-  if (str[0] === "-" || str[0] === "+" || str[0] === ".") {
-    if (str[1] && str[1].match(/\d/)) return false;
-  }
-  const isAlphaNum = ch => ch.match(/[a-zA-Z0-9]/);
-  const isSpecial = ch => ".*+!-_?$%&=<>:#/".includes(ch);
-  return [...str].every(ch => isAlphaNum(ch) || isSpecial(ch));
-}
-
 //------------------------------------------------------------------------------
 // Result
 //------------------------------------------------------------------------------
@@ -126,4 +109,3 @@ export function fixIndent(text) {
   result.text = printResult(result);
   return result;
 }
-