+++ CHANGES.md
@@ -1,5 +1,17 @@
 # CHANGES
 
+## 0.2.0
+
 ## 0.1.0
 
 initial release
+
+- **Zero space**:
+  - Top-level forms
+- **One space**:
+  - Vectors, Maps, Sets
+  - Lists that do NOT start with symbol
+- **Two space**:
+  - Lists starting with symbol, whose second line is NOT _arg aligned_
+- **Arg-aligned**:
+  - Lists starting with symbol, whose second line is _arg aligned_